Added suggested changes
